# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Grinch API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionAuth: []
paths:
...
  /rules:
    get:
      operationId: listRules
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Rule list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createRule
      tags:
        - rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreateRequest'
      responses:
        '201':
          description: Rule created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        default:
          $ref: '#/components/responses/Problem'
  /rules/{id}:
    get:
      operationId: getRule
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Rule detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchRule
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulePatchRequest'
      responses:
        '200':
          description: Rule updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteRule
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Rule deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /rule-targets:
    get:
      operationId: listRuleTargets
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/RuleIdFilter'
        - $ref: '#/components/parameters/SubjectKindFilter'
        - $ref: '#/components/parameters/SubjectIdFilter'
        - $ref: '#/components/parameters/RuleTargetAssignmentFilter'
        - $ref: '#/components/parameters/RulePolicyFilter'
      responses:
        '200':
          description: Rule target list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTargetListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createRuleTarget
      tags:
        - rule-targets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleTargetCreateRequest'
      responses:
        '201':
          description: Rule target created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTarget'
        default:
          $ref: '#/components/responses/Problem'
  /rule-targets/{id}:
    get:
      operationId: getRuleTarget
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Rule target detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTarget'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchRuleTarget
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleTargetPatchRequest'
      responses:
        '200':
          description: Rule target updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTarget'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteRuleTarget
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Rule target deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: grinch_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        format: int32
        minimum: 1
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        format: int32
        minimum: 0
    Search:
      name: search
      in: query
      schema:
        type: string
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Sort field name.
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
...
    RuleIdFilter:
      name: rule_id
      in: query
      schema:
        type: string
        format: uuid
    SubjectIdFilter:
      name: subject_id
      in: query
      schema:
        type: string
        format: uuid
...
    SubjectKindFilter:
      name: subject_kind
      in: query
      schema:
        $ref: '#/components/schemas/RuleTargetSubjectKind'
    RuleTargetAssignmentFilter:
      name: assignment
      in: query
      schema:
        $ref: '#/components/schemas/RuleTargetAssignment'
    RulePolicyFilter:
      name: policy
      in: query
      schema:
        $ref: '#/components/schemas/RulePolicy'
  responses:
    Problem:
      description: Problem response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - code
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        code:
          type: string
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
...
    RuleType:
      type: string
      enum:
        - binary
        - certificate
        - team_id
        - signing_id
        - cd_hash
    RulePolicy:
      type: string
      enum:
        - allowlist
        - blocklist
        - silent_blocklist
        - cel
    RuleTargetAssignment:
      type: string
      enum:
        - include
        - exclude
    RuleTargetSubjectKind:
      type: string
      enum:
        - group
...
    RuleSummary:
      type: object
      required:
        - id
        - name
        - description
        - rule_type
        - identifier
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rule_type:
          $ref: '#/components/schemas/RuleType'
        identifier:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Rule:
      allOf:
        - $ref: '#/components/schemas/RuleSummary'
        - type: object
          required:
            - custom_message
            - custom_url
          properties:
            custom_message:
              type: string
            custom_url:
              type: string
    RuleCreateRequest:
      type: object
      required:
        - name
        - rule_type
        - identifier
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          $ref: '#/components/schemas/RuleType'
        identifier:
          type: string
        custom_message:
          type: string
        custom_url:
          type: string
    RulePatchRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          $ref: '#/components/schemas/RuleType'
        identifier:
          type: string
        custom_message:
          type: string
        custom_url:
          type: string
    RuleTargetSummary:
      type: object
      required:
        - id
        - rule_id
        - subject_kind
        - subject_id
        - assignment
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          format: uuid
        subject_kind:
          $ref: '#/components/schemas/RuleTargetSubjectKind'
        subject_id:
          type: string
          format: uuid
        assignment:
          $ref: '#/components/schemas/RuleTargetAssignment'
        priority:
          type: integer
          format: int32
          nullable: true
        policy:
          $ref: '#/components/schemas/RulePolicy'
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RuleTarget:
      allOf:
        - $ref: '#/components/schemas/RuleTargetSummary'
        - type: object
          properties:
            cel_expression:
              type: string
    RuleTargetCreateRequest:
      type: object
      required:
        - rule_id
        - subject_id
        - assignment
      properties:
        rule_id:
          type: string
          format: uuid
        subject_id:
          type: string
          format: uuid
        assignment:
          $ref: '#/components/schemas/RuleTargetAssignment'
        priority:
          type: integer
          format: int32
        policy:
          $ref: '#/components/schemas/RulePolicy'
        cel_expression:
          type: string
    RuleTargetPatchRequest:
      type: object
      minProperties: 1
      properties:
        subject_id:
          type: string
          format: uuid
        assignment:
          $ref: '#/components/schemas/RuleTargetAssignment'
        priority:
          type: integer
          format: int32
          nullable: true
        policy:
          $ref: '#/components/schemas/RulePolicy'
          nullable: true
        cel_expression:
          type: string
...
    RuleListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/RuleSummary'
    RuleTargetListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/RuleTargetSummary'
...